requirements.txt
Django==4.0.4
djangorestframework==3.13.1
PyMySQL==1.0.2
cryptography==37.0.2